pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-poison-detector"
version = "0.1.0"
description = "Local LLM safety sandbox for prompt-injection and data-poisoning detection"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0,<3.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "torch>=2.4.0",
    "transformers>=4.44.0,<5.0.0",
    "safetensors>=0.4.3",
    "sentencepiece>=0.2.0",
    "huggingface_hub>=0.22.0",
    "accelerate>=0.33.0",
]

[project.scripts]
llmpoison = "run:main"

[project.urls]
Homepage = "https://github.com/Rebas9512/llm-poison-detector"
Repository = "https://github.com/Rebas9512/llm-poison-detector"
Issues = "https://github.com/Rebas9512/llm-poison-detector/issues"

[tool.setuptools]
py-modules = ["run"]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "api*", "static*", "schema*"]

[tool.setuptools.package-data]
"*" = ["static/*.html", "static/*.js", "static/*.css", "schema/*.sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
